+++ source/extensions/lightspeed.trex.utils.widget/docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.8.4]
+### Changed
+- Sourced the Remix categories dialog metadata from the authoritative USD schema.
+- Excluded deprecated Remix categories from new assignment.
+
 ## [1.8.3]
 ### Added
 - Added awaitable message-dialog and invalid project dependency confirmation helpers.